Moved keypair and NVM projects to Core. Moved/Added some helper methods to Core project. More significant changes to transactions. Upped NuGet versions. Added 2 new missing fields of NeoScan balance DTO. Updated Demo project with the refractored methods.

+++ src/NeoModules.RPC/TransactionManagers/ITransactionManager.cs
@@ -10,8 +10,8 @@ namespace NeoModules.RPC.TransactionManagers
     {
         IClient Client { get; set; }
         IAccount Account { get; set; }
-        Task<double> EstimateGasAsync(string serializedScriptHash);
-        Task<double> EstimateGasAsync(string scriptHash, string operation, List<InvokeParameter> parameterList);
+        Task<decimal> EstimateGasAsync(string serializedScriptHash);
+        Task<decimal> EstimateGasAsync(string scriptHash, string operation, List<InvokeParameter> parameterList);
         Task<bool> SendTransactionAsync(string serializedAndSignedTx);
         Task<string> SignTransactionAsync(byte[] transactionData); //TODO change byte[] a specific object
         Task<Transaction> GetTransaction(string tx);
